pyproject.toml
[project]
name = "aikefu-client"
version = "1.0.0"
description = "爱客服采集客户端 - 拼多多消息采集与AI回复Windows桌面应用"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.49.0",
    "websockets>=13.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "PyQt6>=6.7.0",
    "pyqt6-fluent-widgets[full]>=1.6.0",
    "pymysql>=1.1.0",
    "SQLAlchemy>=2.0.0",
    "pycryptodome>=3.20.0",
    "qasync>=0.27.0",
]

[project.optional-dependencies]
build = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
aikefu-client = "app:main"

[tool.setuptools.packages.find]
where = ["."]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"
